requirements.txt
Pillow
PyMuPDF